+++ viteclientts/src/App.tsx
@@ -9,25 +9,35 @@ import CentralizedSnackbar from './components/snackbar/CentralizedSnackbar';
 import ErrorBoundary from './components/ErrorBoundary';
 import CssBaseline from '@mui/material/CssBaseline';
 import NavBar from './components/NavBar';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
 function App() {
+  const queryClient = new QueryClient({
+    defaultOptions: {
+      queries: {
+        retry: 1,
+      },
+    },
+  });
 
   return (
     <SnackbarProvider>
       <ErrorBoundary>
         <CssBaseline />
-        <Router>
-          <div className="App">
-            <NavBar />
-            <Routes>
-              <Route path='/' element={<VillaAnnaHomePage />} />
-              <Route path='/bewaesserung' element={<BewaesserungsPage />} />
-              <Route path='/countdown' element={<CountdownPage />} />
-              <Route path='*' element={<NotFoundPage />} />
-            </Routes>
-            <CentralizedSnackbar />
-          </div>
-        </Router>
+        <QueryClientProvider client={queryClient}>
+          <Router>
+            <div className="App">
+              <NavBar />
+              <Routes>
+                <Route path='/' element={<VillaAnnaHomePage />} />
+                <Route path='/bewaesserung' element={<BewaesserungsPage />} />
+                <Route path='/countdown' element={<CountdownPage />} />
+                <Route path='*' element={<NotFoundPage />} />
+              </Routes>
+              <CentralizedSnackbar />
+            </div>
+          </Router>
+        </QueryClientProvider>
       </ErrorBoundary>
     </SnackbarProvider>
   );